+++ app/javascript/components/activity_picker_slider.js
@@ -3,6 +3,7 @@ const activitiesToReload = () => {
 
   const picker = document.getElementById('intensiveness-pick-activity');
   let intensiveness;
+  let outputCapital;
 
   if (picker) {
     // const activities = document.getElementById('card-activities-all');
@@ -11,10 +12,13 @@ const activitiesToReload = () => {
       event.preventDefault();
       if (picker.value - 1 === 0) {
         intensiveness = 'low';
+        outputCapital = 'Low';
       } else if (picker.value - 2 === 0) {
         intensiveness = 'medium';
+        outputCapital = 'Medium';
       } else {
         intensiveness = 'high';
+        outputCapital = 'High';
       }
       const activeActivity = document.querySelector(".active-activity");
       const activityToDisplay = document.querySelector(`.activity-${intensiveness}`);
@@ -24,32 +28,32 @@ const activitiesToReload = () => {
       activeActivity.classList.remove('active-activity');
       activityToDisplay.classList.remove('hide-activity');
       activityToDisplay.classList.add('active-activity');
-      output.innerHTML = intensiveness;
+      output.innerHTML = outputCapital;
     });
   }
 };
 
 
-const firstInitInPageActivity = () => {
-  const activeSliderInit = document.querySelector(`.activity-medium`);
-  const sliderInit = activeSliderInit.querySelectorAll(".activity-choice");
-  sliderInit[0].classList.add("active-slide");
-  sliderInit[0].classList.remove("hidden-slide");
-  sliderInit[0].classList.add("show");
-};
+// const firstInitInPageActivity = () => {
+//   const activeSliderInit = document.querySelector(`.activity-medium`);
+//   const sliderInit = activeSliderInit.querySelectorAll(".activity-choice");
+//   sliderInit[0].classList.add("active-slide");
+//   sliderInit[0].classList.remove("hidden-slide");
+//   sliderInit[0].classList.add("show");
+// };
 
-const initActivitiesShow = (intensiveness) => {
-  const activeActivity = document.querySelector(".active-activity");
-  const activityToDisplay = document.querySelector(`.activity-${intensiveness}`);
-  const firstActivityToDisplay = activityToDisplay.querySelectorAll(".activity-choice");
-  const firstActiveActivity = activeActivity.querySelectorAll(".activity-choice");
-  firstActivityToDisplay[0].classList.remove('hidden-slide');
-  firstActivityToDisplay[0].classList.add('show');
-  firstActivityToDisplay[0].classList.add('active-slide');
-  firstActiveActivity[0].classList.remove('show');
-  firstActiveActivity[0].classList.add('hidden-slide');
-  firstActiveActivity[0].classList.remove('active-slide');
-};
+// const initActivitiesShow = (intensiveness) => {
+//   const activeActivity = document.querySelector(".active-activity");
+//   const activityToDisplay = document.querySelector(`.activity-${intensiveness}`);
+//   const firstActivityToDisplay = activityToDisplay.querySelectorAll(".activity-choice");
+//   const firstActiveActivity = activeActivity.querySelectorAll(".activity-choice");
+//   firstActivityToDisplay[0].classList.remove('hidden-slide');
+//   firstActivityToDisplay[0].classList.add('show');
+//   firstActivityToDisplay[0].classList.add('active-slide');
+//   firstActiveActivity[0].classList.remove('show');
+//   firstActiveActivity[0].classList.add('hidden-slide');
+//   firstActiveActivity[0].classList.remove('active-slide');
+// };
 
 
-export { activitiesToReload, initActivitiesShow, firstInitInPageActivity };
+export { activitiesToReload };
