+++ app/javascript/packs/application.js
@@ -1,5 +1,4 @@
-require("chartkick")
-require("chart.js")
+
 import "bootstrap";
 import "../plugins/flatpickr";
 import { loadDynamicBannerText } from '../components/banner';
@@ -18,7 +17,7 @@ import { mainActivityMaker } from '../components/activity_master';
 slideToNext();
 slideToPrevious();
 const activities = document.querySelectorAll(".activity-choice");
-if (activities) {
+if (activities.length > 0) {
   selectCardActivity();
   mainActivityMaker();
 }
@@ -56,19 +55,20 @@ if (agendaBullets.length > 0) {
 }
 
 
-import { drawBackgroundColor, drawStacked } from '../plugins/googlecharts';
-const lineChart = document.querySelector(".svg-container1");
-if (lineChart) {
-  drawBackgroundColor();
-}
-const stackchart = document.querySelector(".svg-container2");
-if (stackchart) {
-  drawStacked();
-}
+// import { functionToExport } from '../plugins/googlecharts';
+
+// const lineChart = document.querySelector(".svg-container1");
+// console.log (lineChart);
+// const stackchart = document.querySelector(".svg-container2");
+// console.log(stackchart);
+// if (lineChart) {
+//   drawBackgroundColor();
+// }
+// if (stackchart) {
+//   drawStacked();
+// }
+
+// functionToExport();
 
 // initMapbox();
 
-
-import { activitiesToReload } from '../components/activity_picker_slider'
-
-activitiesToReload();
