+++ app/views/layouts/application.html.erb
@@ -6,6 +6,7 @@
     <meta name="google-signin-client_id" content="client-id.apps.googleusercontent.com">
 
     <script src="https://apis.google.com/js/platform.js" async defer></script>
+    <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
 
     <title>Hife</title>
     <%= csrf_meta_tags %>
@@ -27,6 +28,5 @@
     <script src="../../assets/js/vendor/popper.min.js"></script>
     <script src="../../dist/js/bootstrap.min.js"></script>
  -->
-   <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
   </body>
 </html>
