+++ app/views/meetings/_activities.html.erb
@@ -1,6 +1,51 @@
 <div class="activity-choices">
   <div id="activities" data-ids=<%= activities.map(&:id).to_json %>>
+    <div class="activity-choice fade-in active-slide show">
+
+      <input type="radio" class="hidden hidden-slide" value="<%= activities[0].id %>" name="meeting[activity_ids][]" id="meeting_activity_ids_<%= activities[0].id %>">
+      <label for="meeting_activity_ids_<%= activities[0].id %>" class="hidden-slide">
+        <%= activities[0].name %>
+      </label>
+      <div class="card-activity">
+
+        <div class="card-activity-photo">
+          <%= image_tag(activities[0].photo) %>
+        </div>
+
+        <div class="card-activity-right">
+          <div class="card-activity-infos">
+            <div class="card-activity-title">
+                <%= activities[0].name %>
+                <%= activities[0].intensiveness.capitalize %>
+            </div>
+            <div class="card-activity-title-details">
+              <div>
+                <i class="far fa-clock pr-2"></i></i><%= "#{activities[0].duration} min" %>
+              </div>
+              <div>
+                <i class="fas fa-map-marker-alt pr-2"></i><%= activities[0].address %>
+              </div>
+            </div>
+            <div class="card-activity-description">
+              <%= activities[0].description %>
+            </div>
+          </div>
+          <div class="card-activity-map">
+            <!-- <div class="map" id='<%#= "map_#{activities[0].id}" %>'
+              style="width: 630px; height: 230px;"
+              data-markers-activity="<%#= {lng: activities[0].longitude, lat: activities[0].latitude}.to_json %>"
+              data-markers-company="<%#= {lng: @company.longitude, lat: @company.latitude}.to_json %>"
+              data-mapbox-api-key='<%#= ENV["MAPBOX_API_KEY"] %>'>
+            </div> -->
+          </div>
+        </div>
+      </div>
+
+    </div>
+
+  <% activities.shift %>
   <% activities.each do |activity| %>
+
     <div class="activity-choice fade-in hidden-slide">
 
       <input type="radio" class="hidden hidden-slide" value="<%= activity.id %>" name="meeting[activity_ids][]" id="meeting_activity_ids_<%= activity.id %>">
